feat: IssueList Service 객체 추가
네트워크 통신 로직 분리
